refactor(api): extract rack state controller crate
Move the rack state controller implementation out of the API crate into a new
carbide-rack-controller crate.

Update API setup, config wiring, and rack controller tests to use the extracted
crate while preserving existing behavior. This prepares rack and power-shelf
controller code for further separation from the API crate.
