+++ crates/api/Cargo.toml
@@ -41,6 +41,7 @@ carbide-host-support = { path = "../host-support", default-features = false }
 carbide-mqtt-common = { path = "../mqtt-common" }
 carbide-network = { path = "../network", features = ["sqlx"] }
 carbide-rack = { path = "../rack" }
+carbide-rack-controller = { path = "../rack-controller" }
 carbide-secrets = { path = "../secrets" }
 carbide-version = { path = "../version" }
 carbide-firmware = { path = "../firmware" }
