+++ crates/api/src/state_controller/mod.rs
@@ -19,6 +19,5 @@ pub mod common_services;
 pub(crate) mod external_service_error;
 pub mod machine;
 pub mod power_shelf;
-pub mod rack;
 
 pub use ::state_controller::{controller, db_write_batch, state_change_emitter, state_handler};
