+++ crates/rack-controller/src/created.rs
@@ -15,25 +15,25 @@
  * limitations under the License.
  */
 
-//! Handler for RackState::Discovering.
+//! Handler for RackState::Created.
 //!
-//! The rack waits in Discovering until all machines, switches, and power
-//! shelves that belong to it (via `rack_id` FK) have reached their Ready
-//! state. Once all devices are ready, reprovisioning is triggered and the
-//! rack transitions to Maintenance.
+//! The rack stays in Created until the expected device counts (looked up
+//! from the config-file RackProfile via the rack's `rack_profile_id`
+//! column) match the actual device counts (machines, switches,
+//! power shelves with `rack_id` FK).
 
 use carbide_rack_controller::context::RackStateHandlerContextObjects;
 use carbide_uuid::rack::{RackId, RackProfileId};
 use db::{machine as db_machine, power_shelf as db_power_shelf, switch as db_switch};
 use model::machine::machine_search_config::MachineSearchConfig;
-use model::rack::{FirmwareUpgradeState, RackMaintenanceState, RackState};
+use model::rack::RackState;
 use state_controller::state_handler::{
     StateHandlerContext, StateHandlerError, StateHandlerOutcome,
 };
 
-use crate::state_controller::rack as carbide_rack_controller;
+use crate as carbide_rack_controller;
 
-pub async fn handle_discovering(
+pub async fn handle_created(
     id: &RackId,
     rack_profile_id: Option<&RackProfileId>,
     ctx: &mut StateHandlerContext<'_, RackStateHandlerContextObjects>,
@@ -49,60 +49,55 @@ pub async fn handle_discovering(
 
     let mut txn = ctx.services.db_pool.begin().await?;
 
-    let ready_compute = db_machine::find_machine_ids(
+    let compute_count = db_machine::find_machine_ids(
         txn.as_mut(),
         MachineSearchConfig {
             rack_id: Some(id.clone()),
-            controller_state: Some("ready".into()),
             ..Default::default()
         },
     )
     .await?
     .len() as u32;
-    let ready_switches = db_switch::find_ids(
+    let switch_count = db_switch::find_ids(
         txn.as_mut(),
         model::switch::SwitchSearchFilter {
             rack_id: Some(id.clone()),
-            controller_state: Some("ready".to_string()),
             ..Default::default()
         },
     )
     .await?
     .len() as u32;
-    let ready_shelves = db_power_shelf::find_ids(
+    let power_shelf_count = db_power_shelf::find_ids(
         txn.as_mut(),
         model::power_shelf::PowerShelfSearchFilter {
             rack_id: Some(id.clone()),
-            controller_state: Some("ready".to_string()),
             ..Default::default()
         },
     )
     .await?
     .len() as u32;
 
-    if ready_compute < capabilities.compute.count
-        || ready_switches < capabilities.switch.count
-        || ready_shelves < capabilities.power_shelf.count
+    if compute_count < capabilities.compute.count
+        || switch_count < capabilities.switch.count
+        || power_shelf_count < capabilities.power_shelf.count
     {
         return Ok(StateHandlerOutcome::wait(format!(
-            "waiting for devices ready: compute={}/{}, switch={}/{}, power_shelf={}/{}",
-            ready_compute,
+            "waiting for devices: compute={}/{}, switch={}/{}, power_shelf={}/{}",
+            compute_count,
             capabilities.compute.count,
-            ready_switches,
+            switch_count,
             capabilities.switch.count,
-            ready_shelves,
+            power_shelf_count,
             capabilities.power_shelf.count,
-        ))
-        .with_txn(txn));
+        )));
     }
 
     tracing::info!(
-        "Rack {} all devices ready, transitioning to Maintenance",
-        id
+        "Rack {} has all expected devices (compute={}, switch={}, power_shelf={}). Transitioning to Discovering.",
+        id,
+        compute_count,
+        switch_count,
+        power_shelf_count
     );
-    Ok(StateHandlerOutcome::transition(RackState::Maintenance {
-        maintenance_state: RackMaintenanceState::FirmwareUpgrade {
-            rack_firmware_upgrade: FirmwareUpgradeState::Start,
-        },
-    }))
+    Ok(StateHandlerOutcome::transition(RackState::Discovering).with_txn(txn))
 }
