+++ crates/rack-controller/src/deleting.rs
