+++ crates/rack-controller/src/handler.rs
@@ -31,7 +31,7 @@ use state_controller::state_handler::{
     StateHandler, StateHandlerContext, StateHandlerError, StateHandlerOutcome,
 };
 
-use crate::state_controller::rack as carbide_rack_controller;
+use crate as carbide_rack_controller;
 
 //------------------------------------------------------------------------------
