+++ crates/rack-controller/src/ready.rs
@@ -32,7 +32,7 @@ use state_controller::state_handler::{
     StateHandlerContext, StateHandlerError, StateHandlerOutcome,
 };
 
-use crate::state_controller::rack as carbide_rack_controller;
+use crate as carbide_rack_controller;
 
 const COMPONENT_ERROR_STATE: &str = "error";
 const MACHINE_FAILED_STATE: &str = "failed";
