+++ crates/rack-controller/src/validating.rs
@@ -28,7 +28,7 @@ use state_controller::state_handler::{
     StateHandlerContext, StateHandlerError, StateHandlerOutcome,
 };
 
-use crate::state_controller::rack as carbide_rack_controller;
+use crate as carbide_rack_controller;
 
 //------------------------------------------------------------------------------
 // Helper types
